added API token middleware to schema-server
