+++ schema-server/src/tmp/my-uploads/test.sql
@@ -1,0 +1,3 @@
+CREATE TABLE api.henlohelolo (
+  id serial PRIMARY KEY
+);
